fix some of the merge issues with dani's work

+++ Assets/Scripts/IDestroyable.cs
@@ -3,5 +3,5 @@ using UnityEngine.Events;
 
 interface IDestroyable // for things that can be damaged and destroyed, like players or obstacles. 
 {
-    public void Damage(int damagePoints);
+    public void Damage(float damagePoints);
 }
